Arrangement of architecture: Separating Vue components to Vue, JS and CSS files
