+++ src/css/Cell.css
@@ -1,0 +1,11 @@
+.cell {
+  width: 100px;
+  height: 100px;
+  background-color: black;
+  font-size: 42px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: 3px solid #4AADC1;
+  border-radius: 20px 20px 20px 20px;
+}
