+++ src/css/ScoreBoard.css
@@ -1,0 +1,53 @@
+h1 {
+  font-family: "Arial Black", Gadget, sans-serif;
+  font-size: 26px;
+  letter-spacing: 4.4px;
+  word-spacing: 3.4px;
+  color: #4AADC1;
+  font-weight: 700;
+  text-decoration: none;
+  font-style: normal;
+  font-variant: small-caps;
+  text-transform: uppercase;
+}
+
+p {
+  font-family: "Arial Black", Gadget, sans-serif;
+  font-size: 18px;
+  letter-spacing: 3.2px;
+  word-spacing: 2.4px;
+  color: #4AADC1;
+  font-weight: 400;
+  text-decoration: none solid rgb(68, 68, 68);
+  font-style: normal;
+  font-variant: small-caps;
+  text-transform: none;
+}
+
+.winner {
+  font-family: "Arial Black", Gadget, sans-serif;
+  font-size: 22px;
+  letter-spacing: 3.2px;
+  word-spacing: 2.4px;
+  color: #FFFFFF;
+  font-weight: 700;
+  text-decoration: none solid rgb(68, 68, 68);
+  font-style: normal;
+  font-variant: normal;
+  text-transform: uppercase;
+}
+
+.game-info {
+  flex-wrap: wrap;
+  display: inline;
+  margin: 60vmin 0 0;
+  justify-content: center;
+  align-content: center;
+  font-size: 1.25em;
+  text-align: center;
+  backdrop-filter: blur(10px);
+  background-blend-mode: exclusion;
+  color: white;
+  padding: 30px;
+  /*text-shadow: 5px 1px 3px #7791A4;*/
+}
