+++ src/css/TicTacToeBoard.css
@@ -1,0 +1,8 @@
+.tic-tac-toe-board {
+  flex-wrap: wrap;
+  display: grid;
+  justify-content: center;
+  align-content: center;
+  grid-template-columns: repeat(3, auto);
+  padding: 20px;
+}
