+++ src/css/TicTacToeGame.css
@@ -1,0 +1,19 @@
+body {
+  margin: 0;
+}
+
+.game {
+  background: #000000;
+  background: -moz-linear-gradient(top, #000000 0%, #203542 45%, #2F6084 100%);
+  background: -webkit-linear-gradient(top, #000000 0%, #203542 45%, #2F6084 100%);
+  background: linear-gradient(to bottom, #000000 0%, #203542 45%, #2F6084 100%);
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.game-area {
+  display: flex;
+  flex-flow: column;
+}
