+++ src/js/ClearButton.js
@@ -1,0 +1,10 @@
+export default {
+  name: "clear-button",
+
+  methods: {
+    // Emit event for game Board clearing
+    clearBoard() {
+      this.$emit("clearBoard")
+    }
+  }
+}
